OnlinShop/requirements.txt
asgiref==3.2.10
autopep8==1.4.4
beautifulsoup4==4.9.1
bootstrap4==0.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.1
django-allauth==0.42.0
django-autoslug==1.9.8
django-bootstrap4==2.2.0
django-countries==5.3.3
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-filter==2.3.0
django-import-export==2.3.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
idna==2.8
importlib-metadata==1.7.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.0.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
pep8==1.7.1
Pillow==6.2.2
pycodestyle==2.5.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.17.2
soupsieve==2.0.1
sqlparse==0.2.4
stripe==2.27.0
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.24.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
